autoclip-electron/engine/requirements.txt
yt-dlp>=2023.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
librosa>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
torch>=1.9.0
torchaudio>=0.9.0
tqdm>=4.62.0
json5>=0.9.0
mediapipe>=0.10.0
opencv-python>=4.8.0
pillow>=9.0.0